pytest/requirements.txt
pytest
selenium
requests
Pillow
boto3
pyocr
